moving files and refactor
