--- utils/style.txt
@@ -1,12 +1,0 @@
-xtick.labelsize: 16
-ytick.labelsize: 16
-font.size: 15
-figure.autolayout: True
-figure.figsize: 7.2,4.45
-axes.titlesize : 16
-axes.labelsize : 17
-lines.linewidth : 2
-lines.markersize : 6
-legend.fontsize: 13
-mathtext.fontset: stix
-font.family: STIXGeneral
